add working support for python
Modify setup_python_runstring() to capture stderr to "myCode.pyerr".  Create PY_post_analysis(), which needs more work to record errors/warnings.

Fix re_points to be greedy to allow "}" in the search string.

Add '\n' to all code because Python will not run the last line if the student did not include a final 'enter'.

Modify req_and_prohibit() to change '\r\n\' to '\n\' to better allow searches that simulate '^' and '$'.